+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * Ignore blacklisted frames completely [#235]
 * Create bindings directly inside of a group [#265]
 * Add option to import class-wide abilities per specialization [#267]
+* Add option to prevent toggling an ability off [#248]
 
 ### Fixed
 
@@ -1731,6 +1732,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [#267]: https://github.com/Snakybo/Clicked/pull/267
 [#265]: https://github.com/Snakybo/Clicked/pull/265
 [#263]: https://github.com/Snakybo/Clicked/pull/263
+[#248]: https://github.com/Snakybo/Clicked/pull/248
 [#247]: https://github.com/Snakybo/Clicked/pull/247
 [#245]: https://github.com/Snakybo/Clicked/pull/245
 [#243]: https://github.com/Snakybo/Clicked/issues/243